Add Fused-Attention Layer for AVX2 Platforms (#137)
* add prefer_fp32

* add avx2 mha

* TARGET_512

* follow clang-tidy

* fix ut on linux

* move intrin utils to btla

* move btla templates to mha_dense_wrapper.h

* add f16c target for avx2 intrin

* move tests to mha_dense_tests.cpp

* fix format
