+++ neural_speed/models/llama/llama.cpp
@@ -42,6 +42,9 @@
 #include "models/model_utils/util.h"
 #include "models/models.h"
 
+static const bool NE_ATTN_PREFER_FP32 =
+    getenv("NE_ATTN_PREFER_FP32") != nullptr && std::string("1") == getenv("NE_ATTN_PREFER_FP32");
+
 // evaluate the transformer
 //
 //   - lctx:      model context
@@ -319,6 +322,7 @@ static bool llama_model_eval_internal(model_context* ctx, const model_input* inp
       ne_set_name(V, "V");
 
       ne_attn_flags_t attn_flags = NE_ATTN_FLAG_NONE;
+      if (NE_ATTN_PREFER_FP32) attn_flags |= NE_ATTN_FLAG_PREFER_FP32;
       if (n_total == 0 || !shift_roped_k) attn_flags |= NE_ATTN_FLAG_IS_CAUSAL;  // no causal mask on next-token cases
       struct ne_tensor* KQV_Out = ne_flash_attn(ctx0, Q, K, V, attn_scale, attn_flags);
       struct ne_tensor* KQV_merged_contiguous =
